Adds a method to retrieve the uncompressed rdata for an RR
This is mostly hacked right now abusing the fact that
the root domain can't be compressed and that an empty
packet does not contain any other domains to be used
for compression.

This will not work for RR that consist of more then
one compressed name in their rdata section.

Mostly fixes #73